+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added
 
+* [HexaPDF::Type::Annotations::Square] for rectangle annotations as well as
+  [HexaPDF::Document::Annotations#create_rectangle]
 * [HexaPDF::Type::Annotation::BorderEffect] type class
 * [HexaPDF::Type::Annotations::BorderEffect] module that provides convenience
   access to the border effect dictionary